chore(ci): refactor e2e package and ci tests to speed up ci
