+++ Makefile
@@ -1,7 +1,8 @@
 pkgs := $(shell go list ./...)
 run := .
 count := 1
+timeout := 120m
 
 test: 
-	KNUU_TIMEOUT=120m go test -v $(pkgs) -run $(run) -count=$(count) -timeout 120m
+	KNUU_TIMEOUT=120m go test -v $(pkgs) -run $(run) -count=$(count) -timeout $(timeout)
 .PHONY: test
