+++ e2e/bittwister/bittwister_test.go
@@ -1,4 +1,4 @@
-package basic
+package bittwister
 
 import (
 	"context"
