+++ e2e/system/build_from_git_test.go
@@ -1,4 +1,4 @@
-package basic
+package system
 
 import (
 	"context"
