+++ e2e/system/env_to_json_test.go
@@ -1,4 +1,4 @@
-package basic
+package system
 
 import (
 	"encoding/json"
