+++ e2e/system/external_file_test.go
@@ -1,4 +1,4 @@
-package basic
+package system
 
 import (
 	"io"
