+++ e2e/system/file_test.go
@@ -1,4 +1,4 @@
-package basic
+package system
 
 import (
 	"context"
