+++ e2e/system/file_test_image_cached_test.go
@@ -1,4 +1,4 @@
-package basic
+package system
 
 import (
 	"fmt"
@@ -7,25 +7,24 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
+	"github.com/celestiaorg/knuu/e2e"
 	"github.com/celestiaorg/knuu/pkg/knuu"
 )
 
-func TestFolderCached(t *testing.T) {
+func TestFileCached(t *testing.T) {
 	t.Parallel()
-
 	// Setup
 	executor, err := knuu.NewExecutor()
 	if err != nil {
 		t.Fatalf("Error creating executor: %v", err)
 	}
 
-	// Test logic
 	const numberOfInstances = 10
 	instances := make([]*knuu.Instance, numberOfInstances)
 
 	for i := 0; i < numberOfInstances; i++ {
 		instanceName := fmt.Sprintf("web%d", i+1)
-		instances[i] = assertCreateInstanceNginxWithVolumeOwner(t, instanceName)
+		instances[i] = e2e.AssertCreateInstanceNginxWithVolumeOwner(t, instanceName)
 	}
 
 	var wgFolders sync.WaitGroup
@@ -35,7 +34,7 @@ func TestFolderCached(t *testing.T) {
 			defer wgFolders.Done()
 			instanceName := fmt.Sprintf("web%d", i+1)
 			// adding the folder after the Commit, it will help us to use a cached image.
-			err := instance.AddFolder("resources/html", "/usr/share/nginx/html", "0:0")
+			err = instance.AddFile("resources/html/index.html", "/usr/share/nginx/html/index.html", "0:0")
 			if err != nil {
 				t.Fatalf("Error adding file to '%v': %v", instanceName, err)
 			}
@@ -43,9 +42,9 @@ func TestFolderCached(t *testing.T) {
 	}
 	wgFolders.Wait()
 
-	// Cleanup
 	t.Cleanup(func() {
-		err := assertCleanupInstances(t, executor, instances)
+		// Cleanup
+		err := e2e.AssertCleanupInstances(t, executor, instances)
 		if err != nil {
 			t.Fatalf("Error cleaning up: %v", err)
 		}
