+++ e2e/system/main_test.go
@@ -1,0 +1,19 @@
+package system
+
+import (
+	"os"
+	"testing"
+
+	"github.com/celestiaorg/knuu/pkg/knuu"
+	"github.com/sirupsen/logrus"
+)
+
+func TestMain(m *testing.M) {
+	err := knuu.Initialize()
+	if err != nil {
+		logrus.Fatalf("error initializing knuu: %v", err)
+	}
+	logrus.Infof("Scope: %s", knuu.Scope())
+	exitVal := m.Run()
+	os.Exit(exitVal)
+}
